+++ R/ggKAP.R
@@ -48,6 +48,7 @@
 #' @param legend_label_se a character, the label of the legend that describes the ribbon of parametric confidence interval. Default is computed as function of `ci_level`.
 #' @param legend_label_median a character, the label of the legend that describes the line type of the median survival.
 #' @param legend_label_sim a character, the label of the legend that describes the ribbon of simulation-based confidence interval. Default is computed as function of `ci_level`.
+#' @param guide_legend_nrow a numeric, the number of rows of the color legend. Set a high number if you have many color categories.
 #' @param scale_color_labels passed to `ggplot2::scale_color_manual(labels = )` for the Kaplan-Meier figure, and `ggplot2::scale_y_discrete(labels = )` for the risk table. The default is `waiver()` (calls the default setup of ggplot2), but can accept `scales::label_parsed` if the coloring variable is an expression (`expression()`) and special character or super/subscripts should be shown.
 #' @param scale_color_values passed to `ggplot2::scale_color_manual(values = )` for the Kaplan-Meier figure and the risk table. Default will use the default PMX graphic charter, but named vectors of colors (with names matching with factor levels) are welcome. Levels not available in the data will be dropped.
 #' @param facetting_args a list, arguments that will overwrite the default arguments of `facet_wrap()`. Note that it will not overwrite `vars`, the latter being informed by `facet_var`. Use `list(labeller = ggplot2::label_parsed)` if the facetting variable is an expression (`expression()`) and special character or super/subscripts should be shown.
@@ -109,6 +110,7 @@
 #' # Labelling
 #'
 #' ggKAP(dat, color_var = "AUCQF")
+#' ggKAP(dat, color_var = "AUCQF", guide_legend_nrow = 2) # force the number of rows
 #'
 #' # This will update both the "color legend" label and the y axis of the risk table
 #' ggKAP(dat, color_var = "AUCQF", label_color = "Nice color label")
@@ -179,6 +181,7 @@ ggKAP <- function(data,
                   legend_label_se = paste0(scales::percent(ci_level), " CI"),
                   legend_label_median = "Median Survival",
                   legend_label_sim = paste0("Simulated ", scales::percent(ci_level), " CI"),
+                  guide_legend_nrow = NULL,
                   scale_color_labels = waiver(),
                   scale_color_values = unname(PMXColors_PMXcolors$default),
                   facetting_args = list(),
@@ -405,13 +408,18 @@ ggKAP <- function(data,
     )
 
   # Guides
+  fig <- fig +
+    guides(
+      color = guide_legend(nrow = guide_legend_nrow),
+      fill = guide_legend(nrow = guide_legend_nrow)
+    )
+
   leg <- leg +
     guides(
       linetype = guide_legend(order = 1),
       fill = guide_legend(order = 2),
       shape = guide_legend(order = 3)
     )
-
 
   # Titles
   # labs() don't accept NULL as input, so need to wrap with an if()